chore: move app navigator to root
